smaller fixes

+++ src/main/java/com/summitsync/api/model/GroupTemplate.java
@@ -6,12 +6,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @Table(name = "SS_GroupTemplate")
 public class GroupTemplate extends BaseTemplate {
-    private double pricePerTrainerPerHour;
+    private BigDecimal pricePerTrainerPerHour;
     private int trainerKey;
 }
